+++ src/import_dbf/models.py
@@ -368,26 +368,45 @@ class Dys(models.Model):
     orcid_id = models.TextField(primary_key=True)
     a_n = models.TextField(blank=True, null=True)
     a_w_etatu = models.TextField(blank=True, null=True)
-    a_dysc_1 = models.TextField(blank=True, null=True)
-    a_dysc_2 = models.TextField(blank=True, null=True)
+    a_dysc_1 = models.ForeignKey(
+        "import_dbf.Ldy", on_delete=DO_NOTHING, db_column="a_dysc_1", related_name="+"
+    )
+    a_dysc_2 = models.ForeignKey(
+        "import_dbf.Ldy", on_delete=DO_NOTHING, db_column="a_dysc_2", related_name="+"
+    )
     a_dysc_1_e = models.TextField(blank=True, null=True)
     a_dysc_2_e = models.TextField(blank=True, null=True)
+
     b_n = models.TextField(blank=True, null=True)
     b_w_etatu = models.TextField(blank=True, null=True)
-    b_dysc_1 = models.TextField(blank=True, null=True)
-    b_dysc_2 = models.TextField(blank=True, null=True)
+    b_dysc_1 = models.ForeignKey(
+        "import_dbf.Ldy", on_delete=DO_NOTHING, db_column="b_dysc_1", related_name="+"
+    )
+    b_dysc_2 = models.ForeignKey(
+        "import_dbf.Ldy", on_delete=DO_NOTHING, db_column="b_dysc_2", related_name="+"
+    )
     b_dysc_1_e = models.TextField(blank=True, null=True)
     b_dysc_2_e = models.TextField(blank=True, null=True)
+
     c_n = models.TextField(blank=True, null=True)
     c_w_etatu = models.TextField(blank=True, null=True)
-    c_dysc_1 = models.TextField(blank=True, null=True)
-    c_dysc_2 = models.TextField(blank=True, null=True)
+    c_dysc_1 = models.ForeignKey(
+        "import_dbf.Ldy", on_delete=DO_NOTHING, db_column="c_dysc_1", related_name="+"
+    )
+    c_dysc_2 = models.ForeignKey(
+        "import_dbf.Ldy", on_delete=DO_NOTHING, db_column="c_dysc_2", related_name="+"
+    )
     c_dysc_1_e = models.TextField(blank=True, null=True)
     c_dysc_2_e = models.TextField(blank=True, null=True)
+
     d_n = models.TextField(blank=True, null=True)
     d_w_etatu = models.TextField(blank=True, null=True)
-    d_dysc_1 = models.TextField(blank=True, null=True)
-    d_dysc_2 = models.TextField(blank=True, null=True)
+    d_dysc_1 = models.ForeignKey(
+        "import_dbf.Ldy", on_delete=DO_NOTHING, db_column="d_dysc_1", related_name="+"
+    )
+    d_dysc_2 = models.ForeignKey(
+        "import_dbf.Ldy", on_delete=DO_NOTHING, db_column="d_dysc_2", related_name="+"
+    )
     d_dysc_1_e = models.TextField(blank=True, null=True)
     d_dysc_2_e = models.TextField(blank=True, null=True)